pyproject.toml
[project]
name = "dl-real-time-drilling"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ruff>=0.14.1",
    "polars>=1.6.0",
    "numpy>=1.26.0",
    "mlflow>=2.14.0",
    "onnxruntime>=1.19.0",
    "onnx>=1.16.0",
    "pydantic-settings>=2.5.2",
    "pytorch-lightning>=2.4.0",
    "torch>=2.4.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "fireducks>=1.4.3",
    "ipykernel>=7.0.1",
]
